+++ fastchat/serve/cli.py
@@ -5,176 +5,50 @@ Usage:
 python3 -m fastchat.serve.cli --model ~/model_weights/llama-7b
 """
 import argparse
-import time
-
-import torch
-from transformers import AutoTokenizer, AutoModelForCausalLM, LlamaTokenizer, AutoModel
-
-from fastchat.conversation import conv_templates, SeparatorStyle
-from fastchat.serve.compression import compress_module
-from fastchat.serve.monkey_patch_non_inplace import replace_llama_attn_with_non_inplace_operations
-from fastchat.serve.serve_chatglm import chatglm_generate_stream
+from fastchat.serve.inference import chat_loop, ChatIO
 
 
-def load_model(model_name, device, num_gpus, load_8bit=False, debug=False):
-    if device == "cpu":
-        kwargs = {}
-    elif device == "cuda":
-        kwargs = {"torch_dtype": torch.float16}
-        if num_gpus == "auto":
-            kwargs["device_map"] = "auto"
-        else:
-            num_gpus = int(num_gpus)
-            if num_gpus != 1:
-                kwargs.update({
-                    "device_map": "auto",
-                    "max_memory": {i: "13GiB" for i in range(num_gpus)},
-                })
-    elif device == "mps":
-        kwargs = {"torch_dtype": torch.float16}
-        # Avoid bugs in mps backend by not using in-place operations.
-        replace_llama_attn_with_non_inplace_operations()
-    else:
-        raise ValueError(f"Invalid device: {device}")
+class SimpleChatIO(ChatIO):
+    def prompt_for_input(self, role) -> str:
+        return input(f"{role}: ")
 
-    if "chatglm" in model_name:
-        tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-        model = AutoModel.from_pretrained(model_name, trust_remote_code=True).half().cuda()
-    else:
-        tokenizer = AutoTokenizer.from_pretrained(model_name, use_fast=False)
-        model = AutoModelForCausalLM.from_pretrained(model_name,
-            low_cpu_mem_usage=True, **kwargs)
+    def prompt_for_output(self, role: str):
+        print(f"{role}: ", end="", flush=True)
 
-    if load_8bit:
-        compress_module(model, device)
+    def append_output(self, output: str):
+        print(output, end=" ", flush=True)
 
-    if (device == "cuda" and num_gpus == 1) or device == "mps":
-        model.to(device)
-
-    if debug:
-        print(model)
-
-    return model, tokenizer
+    def finalize_output(self, output: str):
+        print(output, flush=True)
 
 
-@torch.inference_mode()
-def generate_stream(model, tokenizer, params, device,
-                    context_len=2048, stream_interval=2):
-    prompt = params["prompt"]
-    l_prompt = len(prompt)
-    temperature = float(params.get("temperature", 1.0))
-    max_new_tokens = int(params.get("max_new_tokens", 256))
-    stop_str = params.get("stop", None)
+class MarkdownChatIO(ChatIO):
+    def __init__(self):
+        pass
 
-    input_ids = tokenizer(prompt).input_ids
-    output_ids = list(input_ids)
+    def prompt_for_input(self, role) -> str:
+        return input(f"{role}: ")
 
-    max_src_len = context_len - max_new_tokens - 8
-    input_ids = input_ids[-max_src_len:]
+    def prompt_for_output(self, role: str):
+        print(f"{role}: ", end="", flush=True)
 
-    for i in range(max_new_tokens):
-        if i == 0:
-            out = model(
-                torch.as_tensor([input_ids], device=device), use_cache=True)
-            logits = out.logits
-            past_key_values = out.past_key_values
-        else:
-            attention_mask = torch.ones(
-                1, past_key_values[0][0].shape[-2] + 1, device=device)
-            out = model(input_ids=torch.as_tensor([[token]], device=device),
-                        use_cache=True,
-                        attention_mask=attention_mask,
-                        past_key_values=past_key_values)
-            logits = out.logits
-            past_key_values = out.past_key_values
+    def append_output(self, output: str):
+        print(output, end=" ", flush=True)
 
-        last_token_logits = logits[0][-1]
-
-        if device == "mps":
-            # Switch to CPU by avoiding some bugs in mps backend.
-            last_token_logits = last_token_logits.float().to("cpu")
-
-        if temperature < 1e-4:
-            token = int(torch.argmax(last_token_logits))
-        else:
-            probs = torch.softmax(last_token_logits / temperature, dim=-1)
-            token = int(torch.multinomial(probs, num_samples=1))
-
-        output_ids.append(token)
-
-        if token == tokenizer.eos_token_id:
-            stopped = True
-        else:
-            stopped = False
-
-        if i % stream_interval == 0 or i == max_new_tokens - 1 or stopped:
-            output = tokenizer.decode(output_ids, skip_special_tokens=True)
-            pos = output.rfind(stop_str, l_prompt)
-            if pos != -1:
-                output = output[:pos]
-                stopped = True
-            yield output
-
-        if stopped:
-            break
-
-    del past_key_values
+    def finalize_output(self, output: str):
+        print(output, flush=True)
 
 
 def main(args):
-    model_name = args.model_name
-
-    # Model
-    model, tokenizer = load_model(args.model_name, args.device,
-        args.num_gpus, args.load_8bit, args.debug)
-    is_chatglm = "chatglm" in str(type(model)).lower()
-
-    # Chat
-    conv = conv_templates[args.conv_template].copy()
-    while True:
-        try:
-            inp = input(f"{conv.roles[0]}: ")
-        except EOFError:
-            inp = ""
-        if not inp:
-            print("exit...")
-            break
-
-        conv.append_message(conv.roles[0], inp)
-        conv.append_message(conv.roles[1], None)
-
-        if is_chatglm:
-            prompt = conv.messages[conv.offset:]
-            generate_stream_func = chatglm_generate_stream
-            skip_echo_len = len(conv.messages[-2][1]) + 1
-        else:
-            generate_stream_func = generate_stream
-            prompt = conv.get_prompt()
-            skip_echo_len = len(prompt) + 1
-
-        params = {
-            "model": model_name,
-            "prompt": prompt,
-            "temperature": args.temperature,
-            "max_new_tokens": args.max_new_tokens,
-            "stop": conv.sep if conv.sep_style == SeparatorStyle.SINGLE else conv.sep2,
-        }
-
-        print(f"{conv.roles[1]}: ", end="", flush=True)
-        pre = 0
-        for outputs in generate_stream_func(model, tokenizer, params, args.device):
-            outputs = outputs[skip_echo_len:].strip()
-            outputs = outputs.split(" ")
-            now = len(outputs)
-            if now - 1 > pre:
-                print(" ".join(outputs[pre:now-1]), end=" ", flush=True)
-                pre = now - 1
-        print(" ".join(outputs[pre:]), flush=True)
-
-        conv.messages[-1][-1] = " ".join(outputs)
-
-        if args.debug:
-            print("\n", {"prompt": prompt, "outputs": outputs}, "\n")
+    if args.chatio == "simple":
+        chatio = SimpleChatIO()
+    elif args.chatio == "markdown":
+        chatio = MarkdownChatIO()
+    else:
+        raise ValueError(f"Invalid chatio for console: {args.chatio}")
+    chat_loop(args.model_name, args.device, args.num_gpus, args.load_8bit,
+              args.conv_template, args.temperature, args.max_new_tokens,
+              chatio, args.debug)
 
 
 if __name__ == "__main__":
@@ -188,6 +62,8 @@ if __name__ == "__main__":
         help="Conversation prompt template.")
     parser.add_argument("--temperature", type=float, default=0.7)
     parser.add_argument("--max-new-tokens", type=int, default=512)
+    parser.add_argument("--chatio", type=str, default="markdown",
+                        choices=["simple", "markdown"], help="Chat IO format.")
     parser.add_argument("--debug", action="store_true")
     args = parser.parse_args()
     main(args)
